backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

PyPDF2==3.0.1
nltk==3.8.1

openai>=1.40.0